// apps/web/generated/api/models/systemAnalyticsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerStatisticsResponse } from './customerStatisticsResponse';
import type { DashboardStatisticsResponse } from './dashboardStatisticsResponse';
import type { FreelancerStatisticsResponse } from './freelancerStatisticsResponse';
import type { ProjectStatisticsResponse } from './projectStatisticsResponse';
import type { UserStatisticsResponse } from './userStatisticsResponse';

export interface SystemAnalyticsResponse {
  dashboard: DashboardStatisticsResponse;
  users: UserStatisticsResponse;
  projects: ProjectStatisticsResponse;
  freelancers: FreelancerStatisticsResponse;
  customers: CustomerStatisticsResponse;
}